Added Scroll to Top Component to Home, Teams, and Blog
